# openapi.yaml
openapi: '3.0.0'
info:
  title: Priority ERP Mock API
  version: '1.0.0'
  description: Mock Priority ERP OData-style API for workshop on building AI agents with Google Gemini SDK
servers:
  - url: http://localhost:3000
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Order:
      type: object
      properties:
        ORDNAME:
          type: string
        CUSTNAME:
          type: string
        CURDATE:
          type: string
          format: date
        ORDSTATUS:
          type: string
          enum: [O, C, P]
        QPRICE:
          type: number
        DETAILS:
          type: string
        ORDERITEMS_SUBFORM:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    OrderItem:
      type: object
      properties:
        LINE:
          type: integer
          description: Auto-generated line number
        PARTNAME:
          type: string
        TQUANT:
          type: number
        PRICE:
          type: number
        DUEDATE:
          type: string
          format: date
        KLINE:
          type: string
...
    Supplier:
      type: object
      properties:
        SUPNAME:
          type: string
        SUPDES:
          type: string
        PHONE:
          type: string
        EMAIL:
          type: string
        ADDRESS:
          type: string
        COUNTRY:
          type: string
        STATUS:
          type: string
    SupplierList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Supplier'
...
    ProductionPlan:
      type: object
      properties:
        PLANNAME:
          type: string
        PLANDES:
          type: string
        STARTDATE:
          type: string
          format: date
        ENDDATE:
          type: string
          format: date
        STATUS:
          type: string
    ProductionPlanList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/ProductionPlan'
...
    Timesheet:
      type: object
      properties:
        TSNUM:
          type: string
        EMPNUM:
          type: string
        TSDATE:
          type: string
          format: date
        HOURS:
          type: number
        PROJECT:
          type: string
        DESCRIPTION:
          type: string
    TimesheetList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Timesheet'
...
  parameters:
    filterParam:
      name: $filter
      in: query
      description: OData filter expression
      schema:
        type: string
    orderbyParam:
      name: $orderby
      in: query
      description: OData orderby expression
      schema:
        type: string
    topParam:
      name: $top
      in: query
      description: Number of records to return
      schema:
        type: integer
    skipParam:
      name: $skip
      in: query
      description: Number of records to skip
      schema:
        type: integer
    selectParam:
      name: $select
      in: query
      description: Comma-separated list of fields to return
      schema:
        type: string
    expandParam:
      name: $expand
      in: query
      description: Subform to expand inline
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/ORDERS:
    get:
      summary: List orders
      operationId: listOrders
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
        - $ref: '#/components/parameters/expandParam'
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create an order
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/ORDERS('{ORDNAME}')":
    get:
      summary: Get an order by key
      operationId: getOrder
      tags: [Orders]
      parameters:
        - name: ORDNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update an order
      operationId: updateOrder
      tags: [Orders]
      parameters:
        - name: ORDNAME
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an order
      operationId: deleteOrder
      tags: [Orders]
      parameters:
        - name: ORDNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Order deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/SUPPLIERS:
    get:
      summary: List suppliers
      operationId: listSuppliers
      tags: [Suppliers]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of suppliers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a supplier
      operationId: createSupplier
      tags: [Suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Supplier'
      responses:
        '201':
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/SUPPLIERS('{SUPNAME}')":
    get:
      summary: Get a supplier by key
      operationId: getSupplier
      tags: [Suppliers]
      parameters:
        - name: SUPNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a supplier
      operationId: updateSupplier
      tags: [Suppliers]
      parameters:
        - name: SUPNAME
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Supplier'
      responses:
        '200':
          description: Supplier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a supplier
      operationId: deleteSupplier
      tags: [Suppliers]
      parameters:
        - name: SUPNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Supplier deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== WORK ORDERS ====================
...
  /api/PRODUCTIONPLANS:
    get:
      summary: List production plans
      operationId: listProductionPlans
      tags: [Manufacturing]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of production plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionPlanList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a production plan
      operationId: createProductionPlan
      tags: [Manufacturing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionPlan'
      responses:
        '201':
          description: Production plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/PRODUCTIONPLANS('{PLANNAME}')":
    get:
      summary: Get a production plan by key
      operationId: getProductionPlan
      tags: [Manufacturing]
      parameters:
        - name: PLANNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Production plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a production plan
      operationId: updateProductionPlan
      tags: [Manufacturing]
      parameters:
        - name: PLANNAME
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionPlan'
      responses:
        '200':
          description: Production plan updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a production plan
      operationId: deleteProductionPlan
      tags: [Manufacturing]
      parameters:
        - name: PLANNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Production plan deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== EMPLOYEES ====================
...
  /api/TIMESHEETS:
    get:
      summary: List timesheets
      operationId: listTimesheets
      tags: [HR]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of timesheets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a timesheet entry
      operationId: createTimesheet
      tags: [HR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Timesheet'
      responses:
        '201':
          description: Timesheet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timesheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/TIMESHEETS('{TSNUM}')":
    get:
      summary: Get a timesheet by key
      operationId: getTimesheet
      tags: [HR]
      parameters:
        - name: TSNUM
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Timesheet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timesheet'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a timesheet
      operationId: updateTimesheet
      tags: [HR]
      parameters:
        - name: TSNUM
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Timesheet'
      responses:
        '200':
          description: Timesheet updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timesheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a timesheet
      operationId: deleteTimesheet
      tags: [HR]
      parameters:
        - name: TSNUM
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Timesheet deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== DEPARTMENTS ====================
...